Tetris Python code Project
Tetris is one of the most iconic and widely recognized video games in history. It was originally designed and programmed by Alexey Pajitnov, a Soviet computer scientist, in 1984. Since then, it has become a global phenomenon, influencing gaming culture and the development of puzzle games. The game’s simplicity and addictiveness have made it a classic that remains popular across various platforms.